Solution to Find and Display the Top Scorer

+++ Find and Display the Top Scorer@@67877d61dc107486ad5c2744/code.c
@@ -7,7 +7,7 @@ void main(){
     char name[10];
     float marks;
     scanf("%d",&len);
-    int max=-99;
+    float max=-99;
     int mrno;
     char mname[10];
